long to int

+++ PeerPilot/viewquizresults/viewquizresults.cpp
@@ -114,7 +114,7 @@ void viewquizresults::on_markGradePushButton_clicked(){
     bool ok = false;
     int grade = QInputDialog::getInt(this,"Mark question as graded?",
                                      "Enter the maximum possible score for this question:\n" + ui->questionComboBox->currentText(),
-                                     0,0, std::numeric_limits<long>::max(),1, &ok);
+                                     0,0, std::numeric_limits<int>::max(),1, &ok);
 
     if(!ok){
         return;
@@ -152,7 +152,7 @@ void viewquizresults::on_exportGradesPushButton_clicked(){
 
     int maxPoints = QInputDialog::getInt(this,"Exporting gradesheet",
                                          "Enter the maximum points possible for this assignment:\n",
-                                         0,0,std::numeric_limits<long>::max(), 1, &ok);
+                                         0,0,std::numeric_limits<int>::max(), 1, &ok);
 
     if(!ok){
         return;
